# examples/paygate-rust-nuxt/spec/openapi/openapi.yaml
openapi: "3.1.0"
info:
  title: paygate API
  version: "1.0.0"
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    paygate is a **payment service**: merchants integrate with it, and it
    integrates ECPay and NewebPay on their behalf as a **platform
    merchant** (ECPay's own term is contracted partner platform merchant). Each merchant keeps its own account at the
    provider; paygate holds the platform credentials that sign for all of them,
    and the money is settled to the merchant's own account, never to paygate's.
    This file is the surface merchants and customers see. What paygate sends the
    providers is in `payment-provider.yaml`; what it sends merchants is in
    `demo-merchant.yaml`.

    ## The steps, and where each credential lives
    | Step | Who calls | Credential |
    |------|-----------|------------|
    | `POST /payments` — create the order AND get the provider's signed form | merchant server | `Authorization: Bearer sk_test_…` |
    | `POST /webhooks/{provider}` — the outcome | the PROVIDER | the provider's signature in the body |
    | the callback to `notify_url` | paygate → merchant | `CheckMacValue` in the body, which the merchant verifies |

    **There is no endpoint here that a customer's browser calls.** The merchant's
    own page renders the form this API returns and submits it to the provider, so
    the customer talks to the merchant and to the provider and never to paygate.
...

...

...

...

...

...

    ## Money
...

...

...
servers:
...
paths:
  /payments:
    post:
      operationId: createOrder
      summary: Create an order for a customer to pay
      tags: [Orders]
      security: [ { MerchantApiKey: [] } ]
      description: |
        The merchant's server creates the order and gets back somewhere to send
        the customer. No card data is involved and no provider is called: which
        provider will take the payment is decided later, when the customer
        actually goes to check out.

        Order of checks, each before the next: API key (`401`), rate limit
        (`429`), `Idempotency-Key` presence (`400`), idempotency lookup (replay
        / `409` / `422`), body shape (`400 INVALID_REQUEST` — an unknown field,
        a wrong type, or anything that looks like card data), field rules
        (`422 VALIDATION_FAILED`), currency (`422 CURRENCY_NOT_SUPPORTED`),
        trade number (`422 MERCHANT_TRADE_NO_TAKEN`).

        The order, its `PaymentCreated` audit row and the idempotency record are
        committed together with the **attempt** the form belongs to and its
        `PaymentAttemptStarted` row, and then the form is signed.

        Calling it again for an unpaid order with the same parameters gets the
        same order and a NEW `provider_trade_no`, because ECPay refuses one it has
        seen before. A retry carrying the same `Idempotency-Key` is replayed
        instead, form and all, and opens nothing.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateOrderRequest" }
      responses:
        "201":
          description: Created, awaiting payment
          headers:
            Idempotent-Replayed: { $ref: "#/components/headers/IdempotentReplayed" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreatedOrder" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "409": { $ref: "#/components/responses/IdempotencyKeyInUse" }
        "422":
          description: >
            `VALIDATION_FAILED` (with `field`), `CURRENCY_NOT_SUPPORTED`,
            `MERCHANT_TRADE_NO_TAKEN`, or `IDEMPOTENCY_KEY_REUSED`.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ValidationError"
                  - $ref: "#/components/schemas/Error"
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/Internal" }

    get:
      operationId: findOrderByTradeNo
      summary: Find one of the caller's orders by the merchant's own number
      tags: [Orders]
      security: [ { MerchantApiKey: [] } ]
      parameters:
        - name: merchant_trade_no
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: The order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Payment" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/PaymentNotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
components:
  securitySchemes:
    MerchantApiKey:
      type: http
      scheme: bearer
      description: A secret key issued in the dashboard.
    ProviderCallbackSignature:
      type: apiKey
      in: query
      name: CheckMacValue
      description: >
        Carried as a FORM FIELD, which OpenAPI has no scheme for, so this
        declaration is documentation. ECPay sends `CheckMacValue` (SHA-256 over
        the sorted parameters wrapped in the HashKey and HashIV the provider
        issued paygate); NewebPay sends `TradeSha` over the encrypted `TradeInfo`.
        paygate recomputes it and compares in constant time before anything is
        parsed or applied.
    DashboardSession:
      type: apiKey
      in: cookie
      name: session
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 1, maxLength: 255 }
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      schema: { type: string, maxLength: 128 }
...
  headers:
    IdempotentReplayed:
      description: "`true` when the body is a stored answer being replayed."
      schema: { type: string, const: "true" }
    RetryAfter:
      description: Whole seconds until the merchant's bucket holds one token.
      schema: { type: integer, minimum: 1 }
  responses:
    BadRequest:
      description: >
        `IDEMPOTENCY_KEY_MISSING`, `IDEMPOTENCY_KEY_INVALID`, or
        `INVALID_REQUEST` — malformed JSON, an unknown field, a wrong type, or
        card data sent to an endpoint that must not receive it.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthenticated:
      description: No valid credential of the kind this endpoint accepts
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    PaymentNotFound:
      description: No such order for the caller
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    IdempotencyKeyInUse:
      description: A request with this key is still being processed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    RateLimited:
      description: The merchant's rate limit is exhausted
      headers:
        Retry-After: { $ref: "#/components/headers/RetryAfter" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Internal:
      description: >
        `INTERNAL`. Nothing a caller can fix, and nothing this gateway will
        guess at: the one reachable case is an idempotency record whose stored
        answer can no longer be replayed faithfully — a status that is not a
        status — which is answered as the failure it is rather than replayed as
        a success (spec.md, *Two kinds of duplicate*).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    CreateOrderRequest:
      type: object
      additionalProperties: false
      required: [merchant_trade_no, amount, currency, item_desc, notify_url, client_back_url]
      description: >
        No card fields, and `additionalProperties: false` so that an
        integration which tries to send one is told at the first attempt rather
        than having it quietly dropped.
      properties:
        merchant_trade_no:
          type: string
          pattern: "^[A-Za-z0-9-]{1,20}$"
          description: The merchant's own order number, unique per merchant (ECPay's MerchantTradeNo).
        amount: { type: integer, minimum: 1, maximum: 99999999 }
        currency: { type: string, pattern: "^[A-Z]{3}$" }
        item_desc: { type: string, minLength: 1, maxLength: 200 }
        notify_url:
          type: string
          pattern: "^/[^\\s]*$"
          maxLength: 200
          description: >
            Where paygate posts the signed result, server to server — the
            merchant's own ReturnURL, one level down. In this example every
            service sits behind one reverse proxy, so it is an absolute PATH; a
            real deployment would take an absolute `https` URL and apply ECPay's
            own rules to it (a public host, not a CDN address, ports 80 or 443,
            punycode for a non-ASCII domain).
        client_back_url:
          type: string
          pattern: "^/[^\\s]*$"
          maxLength: 200
          description: >
            Where the PROVIDER sends the customer when they are done — paygate
            passes this straight through as `ClientBackURL` and never receives
            the customer itself. The 200-character limit is not paygate's choice:
            it is ECPay's limit on that parameter, inherited by whoever fills it.
    CreatedOrder:
      allOf:
        - $ref: "#/components/schemas/Payment"
        - $ref: "#/components/schemas/ProviderHandOff"
    ProviderHandOff:
      type: object
      required: [provider, action, fields]
      description: >
        The form the merchant's page must submit, and the only reason this call
        exists. ECPay's own plugin builds this locally and makes no call at all,
        because that merchant holds its own `HashKey`; under a platform the key is
        paygate's, so signing is the one thing a merchant cannot do for itself.
      properties:
        provider:
          type: string
          enum: [ecpay, newebpay]
          description: Which gateway this order was routed to.
        action:
          type: string
          description: >
            The provider's cashier URL — another origin, which is the point. The
            merchant's page posts the fields here and the customer leaves.
          example: https://payment.ecpay.com.tw/Cashier/AioCheckOut/V5
        fields:
          type: object
          additionalProperties: { type: string }
          description: >
            Every parameter that provider expects, signature included, exactly as
            it must be sent — ECPay's set in the clear (`MerchantID`, `PlatformID`,
            `MerchantTradeNo`, `TotalAmount`, `ChoosePayment`, `CheckMacValue`
            and the rest), NewebPay's as `MerchantID`, `TradeInfo`, `TradeSha`,
            `Version`. paygate's key is in neither, and nor is
            `OrderResultURL`: `ClientBackURL` is the merchant's own page, which
            is where the customer goes back to.
    Payment:
      type: object
      required: [id, merchant_trade_no, status, amount, currency, item_desc, amount_refunded, created_at]
      properties:
        id: { type: string, format: uuid, description: "paygate's trade number (ECPay's TradeNo)" }
        merchant_trade_no: { type: string }
        status:
          type: string
          enum: [pending, succeeded, refunded]
          description: >
            Three states, and neither `processing` nor `failed` is one of them.
            paygate cannot know whether a customer is at the provider or has
            closed the tab, so an order with forms outstanding is still
            `pending`; what is outstanding is on `last_attempt`, where it is a
            fact rather than a guess (spec.md, "State").
        amount: { type: integer }
        currency: { type: string }
        item_desc: { type: string }
        amount_refunded: { type: integer }
        card_brand: { type: [string, "null"], enum: [visa, mastercard, amex, unknown, null] }
        card_last4: { type: [string, "null"], pattern: "^[0-9]{4}$" }
        failure_code: { type: [string, "null"] }
        client_back_url: { type: string }
        last_attempt:
          oneOf:
            - type: "null"
            - type: object
              required: [provider, status]
              properties:
                provider: { type: string, enum: [ecpay, newebpay] }
                status: { type: string, enum: [redirected, succeeded, failed, abandoned] }
                failure_code: { type: [string, "null"] }
          description: >
            What became of the most recent form. `redirected` is the window in
            which the customer is at the provider — paying, authenticating, or
            gone — and nobody here knows the answer yet. It is the honest
            substitute for a `processing` status, and a merchant that wants to
            show "we are waiting" reads this.
        created_at: { type: string, format: date-time }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          enum:
            - INVALID_REQUEST
            - IDEMPOTENCY_KEY_MISSING
            - IDEMPOTENCY_KEY_INVALID
            - IDEMPOTENCY_KEY_IN_USE
            - IDEMPOTENCY_KEY_REUSED
            - MERCHANT_TRADE_NO_TAKEN
            - UNAUTHENTICATED
            - INVALID_CREDENTIALS
            - PAYMENT_NOT_FOUND
            - API_KEY_NOT_FOUND
            - CURRENCY_NOT_SUPPORTED
            - PAYMENT_NOT_REFUNDABLE
            - REFUND_EXCEEDS_REMAINING
            - RANGE_TOO_LARGE
            - RATE_LIMITED
            - PROVIDER_UNAVAILABLE
            - SESSION_STORE_UNAVAILABLE
            - REPORTING_UNAVAILABLE
            - INTERNAL
    ValidationError:
      type: object
      required: [error, field]
      properties:
        error: { type: string, const: VALIDATION_FAILED }
        field: { type: string }